[planning]: refactor the joint_traj pub file name; separate the pub_mpc_pred_xu as an individual file
